+++ dio/test/test_suite_test.dart
@@ -3,6 +3,6 @@ import 'package:dio_test/tests.dart';
 
 void main() {
   dioAdapterTestSuite(
-    () => Dio(BaseOptions(baseUrl: 'https://httpbun.com/')),
+    (baseUrl) => Dio(BaseOptions(baseUrl: baseUrl)),
   );
 }