+++ dio/test/timeout_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:dio/io.dart';
+import 'package:dio_test/util.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -9,7 +10,7 @@ void main() {
 
   setUp(() {
     dio = Dio();
-    dio.options.baseUrl = 'https://httpbun.com/';
+    dio.options.baseUrl = httpbunBaseUrl;
   });
 
   group('Timeout exception of', () {
@@ -17,7 +18,7 @@ void main() {
       test('update between calls', () async {
         final client = HttpClient();
         final dio = Dio()
-          ..options.baseUrl = 'https://httpbun.com'
+          ..options.baseUrl = nonRoutableUrl
           ..httpClientAdapter = IOHttpClientAdapter(
             createHttpClient: () => client,
           );