+++ dio_test/lib/src/test/basic_tests.dart
@@ -1,0 +1,81 @@
+import 'dart:io';
+
+import 'package:dio/dio.dart';
+import 'package:dio_test/util.dart';
+import 'package:test/test.dart';
+
+void basicTests(
+  Dio Function(String baseUrl) create,
+) {
+  late Dio dio;
+
+  setUp(() {
+    dio = create(httpbunBaseUrl);
+  });
+
+  group('basic request', () {
+    test(
+      'works with non-TLS requests',
+      () async {
+        await dio.get('http://flutter.cn/');
+        await dio.get('https://flutter.cn/non-exist-destination');
+      },
+      testOn: 'vm',
+    );
+
+    test('fails with an invalid HTTP URL', () {
+      expectLater(
+        dio.get('http://does.not.exist'),
+        throwsDioException(
+          DioExceptionType.connectionError,
+          matcher: kIsWeb
+              ? null
+              : isA<DioException>().having(
+                  (e) => e.error,
+                  'inner exception',
+                  isA<SocketException>(),
+                ),
+        ),
+      );
+    });
+
+    test('fails with an invalid HTTPS URL', () {
+      expectLater(
+        dio.get('https://does.not.exist'),
+        throwsDioException(
+          DioExceptionType.connectionError,
+          matcher: kIsWeb
+              ? null
+              : isA<DioException>().having(
+                  (e) => e.error,
+                  'inner exception',
+                  isA<SocketException>(),
+                ),
+        ),
+      );
+    });
+
+    test('throws DioException that can be caught', () async {
+      try {
+        await dio.get('https://does.not.exist');
+        fail('did not throw');
+      } on DioException catch (e) {
+        expect(e, isNotNull);
+      }
+    });
+
+    test('POST string', () async {
+      final response = await dio.post('/post', data: 'TEST');
+      expect(response.data['data'], 'TEST');
+    });
+
+    test('POST map', () async {
+      final response = await dio.post(
+        '/post',
+        data: {'a': 1, 'b': 2, 'c': 3},
+      );
+      expect(response.data['data'], '{"a":1,"b":2,"c":3}');
+      expect(response.statusCode, 200);
+    });
+  });
+}