+++ dio_test/lib/src/test/redirect_tests.dart
@@ -1,22 +1,21 @@
 import 'package:dio/dio.dart';
+import 'package:dio_test/util.dart';
 import 'package:test/test.dart';
 
-import '../utils.dart';
-
 void redirectTests(
-  Dio Function() create,
+  Dio Function(String baseUrl) create,
 ) {
   late Dio dio;
 
-  setUpAll(() {
-    dio = create();
+  setUp(() {
+    dio = create(httpbunBaseUrl);
   });
 
   group('redirects', () {
     test('single', () async {
       final response = await dio.get(
         '/redirect',
-        queryParameters: {'url': 'https://httpbun.com/get'},
+        queryParameters: {'url': '$httpbunBaseUrl/get'},
         onReceiveProgress: (received, total) {
           // ignore progress
         },
@@ -48,6 +47,28 @@ void redirectTests(
         expect(ri.statusCode, 302);
         expect(ri.method, 'GET');
       }
+    });
+
+    test(
+      'empty location',
+      () async {
+        final response = await dio.get(
+          '/redirect',
+        );
+        expect(response.isRedirect, isTrue);
+        expect(response.redirects.length, 1);
+
+        final ri = response.redirects.first;
+        expect(ri.statusCode, 302);
+        expect(ri.location.path, '/get');
+        expect(ri.method, 'GET');
+      },
+      skip: 'Httpbun does not support empty location redirects',
+    );
+
+    test('request with redirect', () async {
+      final res = await dio.get('/absolute-redirect/2');
+      expect(res.statusCode, 200);
     });
   });
 }