+++ dio_test/lib/src/test/suite.dart
@@ -2,22 +2,26 @@ import 'package:dio/dio.dart';
 import 'package:dio_test/tests.dart';
 
 typedef TestSuiteFunction = void Function(
-  Dio Function() create,
+  Dio Function(String baseUrl) create,
 );
 
 const _tests = [
+  basicTests,
+  cancellationTests,
   corsTests,
-  downloadStreamTests,
+  downloadTests,
   headerTests,
   httpMethodTests,
   parameterTests,
   redirectTests,
   statusCodeTests,
   timeoutTests,
+  uploadTests,
+  urlEncodedTests,
 ];
 
 void dioAdapterTestSuite(
-  Dio Function() create, {
+  Dio Function(String baseUrl) create, {
   List<TestSuiteFunction> tests = _tests,
 }) =>
     tests.forEach((test) => test(create));