+++ plugins/http2_adapter/test/test_suite_test.dart
@@ -4,7 +4,7 @@ import 'package:dio_test/tests.dart';
 
 void main() {
   dioAdapterTestSuite(
-    () => Dio(BaseOptions(baseUrl: 'https://httpbun.com/'))
+    (baseUrl) => Dio(BaseOptions(baseUrl: baseUrl))
       ..httpClientAdapter = Http2Adapter(ConnectionManager()),
   );
 }