+++ dio/CHANGELOG.md
@@ -5,6 +5,14 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 ## Unreleased
 
+- Remove sockets detach in `IOHttpClientAdapter`.
+
+## 5.4.2+1
+
+- Revert "Catch sync/async exceptions in interceptors' handlers".
+
+## 5.4.2
+
 - Fix `receiveTimeout` throws exception after the request has been cancelled.
 - Catch sync/async exceptions in interceptors' handlers.
 - Throws precise `StateError` for handler's duplicated calls.